backend/requirements.txt
streamlit==1.34.0
langchain==0.1.20
langchain-community==0.0.38
langchain-google-vertexai==1.0.3
pytube==15.0.0
youtube-transcript-api==0.6.2
pypdf==4.2.0
tiktoken==0.7.0
uvicorn==0.29.0
fastapi==0.111.0
tqdm==4.66.4
